Change output format to be compatible with input format
This allows you to read from one EFM8 and flash onto another based on
the "Read" tabs output. I've done this a couple of times to read the
firmware from an older device I had and then was able to upload it to
a newer device. Reading the data back again yielded the exact same data.